update Firebase configuration, enhance Device model with JSON serialization, and implement device management in ViewModel

+++ eco_tracker/lib/views/devices/devices_view.dart
@@ -114,7 +114,7 @@ class DevicesView extends GeneralPage {
                               model: modelController.text,
                               manufacturer: manufacturerController.text,
                               category: categoryController.text,
-                              powerConsumption: double.parse(
+                              powerConsumption: int.parse(
                                 powerConsumptionController.text,
                               ),
                             );
